Refactorying.
